add changes of nikolai

+++ MvcLayer/Controllers/ScopeWorksController.cs
@@ -11,6 +11,7 @@ using DatabaseLayer.Models.KDO;
 using OfficeOpenXml.ConditionalFormatting;
 using System.Diagnostics.Contracts;
 using NuGet.Packaging;
+using BusinessLayer.Enums;
 
 namespace MvcLayer.Controllers
 {
@@ -56,7 +57,7 @@ namespace MvcLayer.Controllers
         {
             var obj = _contractService.GetById(contractId);
             int? scopeId;
-            if (!_contractService.IsThereScopeWorks(obj.Id, out scopeId))
+            if (!_contractService.IsThereScopeWorks(contractId, false, out scopeId))
             {
                 TempData["Message"] = "Не заполнены объемы работ!";
                 var urlReturn = returnContractId == 0 ? contractId : returnContractId;
@@ -250,55 +251,46 @@ namespace MvcLayer.Controllers
         {
             ViewData["returnContractId"] = returnContractId;
 
-            //проверяем, к подобъекту этот объем относиться или нет
             if (scopeWork is not null)
             {
-                int mainContractId = 0;
-                var isNotGenContract = _contractService.IsNotGenContract(scopeWork?.ContractId, out mainContractId);
+                int parentContrId = 0;
                 var contract = scopeWork?.ContractId is not null ? _contractService.GetById((int)scopeWork?.ContractId) : null;
+                var contractType = _contractService.GetContractType(contract, out parentContrId);
+                int operationSign = ((contractType == ContractType.GenСontract) || (contractType == ContractType.MultipleContract)) == true ? 1 : -1;
 
-                ///если без ДС
-                if (contract is not null && scopeWork?.IsChange != true)
+                if (contract is not null)
                 {
                     var newScpId = _scopeWork.Create(_mapper.Map<ScopeWorkDTO>(scopeWork));
-                    //если генконтракт, присоздании впервый раз, создаем и объем собст. силами, с теми же значениями (далее их просто обновлять будем)
-                    if (!isNotGenContract && newScpId.HasValue)
-                    {
-                        if (_scopeWork.GetById((int)newScpId) is not null)
-                        {
-                            scopeWork.IsOwnForces = true;
-                            _scopeWork.Create(_mapper.Map<ScopeWorkDTO>(scopeWork));
-                        }
-                    }
-                    if (isNotGenContract)
-                    {
-                        if (contract?.IsOneOfMultiple ?? false)
-                        {
-                            CreateOrUpdateSWCostsOfMainContract(mainContractId, false, scopeWork.SWCosts);
-                            CreateOrUpdateSWCostsOfMainContract(mainContractId, true, scopeWork.SWCosts);
-                        }
-                        else
-                        {
-                            _scopeWork.AddOrSubstractCostsOwnForceMnContract(mainContractId, scopeWork.SWCosts, -1);
-                        }
-                    }
-                }
 
-                ///если по ДС
-                if (contract is not null && scopeWork?.IsChange == true)
-                {
-                    var scopeWorkId = (int)_scopeWork.Create(_mapper.Map<ScopeWorkDTO>(scopeWork));
-                    //проверка создается объем  работ с изменениями или нет
                     if (scopeWork?.AmendmentId is not null && scopeWork?.AmendmentId > 0)
                     {
-                        _scopeWork.AddAmendmentToScopeWork((int)scopeWork?.AmendmentId, scopeWorkId);
+                        _scopeWork.AddAmendmentToScopeWork((int)scopeWork?.AmendmentId, (int)newScpId);
                     }
-                    if (isNotGenContract)
+
+                    //comments
+                    /*
+                       1) если генконтракт -> создаем дополнительно объем собст. силами
+                       2) если подобъект -> создаем дополнительно объем собст. силами, а также обновляем данные генконтракта (объем, объем соб.силами)
+                       3) если субподряд или соглашение -> проверяем на наличие родительского договора (подобъект или генконтракт), и обновляем их объемы соб.силами (вычетаем значения)
+                    */
+
+                    if ((contractType == ContractType.GenСontract) || (contractType == ContractType.MultipleContract) && newScpId.HasValue)
                     {
-                        UpdateSWCostsOfMainContract(mainContractId, (int)scopeWork?.ChangeScopeWorkId, false, scopeWork.SWCosts);
-                        if ((bool)contract?.IsOneOfMultiple)
+                        _scopeWork.AddOwnForcesCostsByScopeId(_mapper.Map<ScopeWorkDTO>(scopeWork), operationSign);
+
+                        if (contractType == ContractType.MultipleContract)
                         {
-                            UpdateSWCostsOfMainContract(mainContractId, (int)scopeWork?.ChangeScopeWorkId, true, scopeWork.SWCosts);
+                            _scopeWork.UpdateParentCosts(parentContrId, scopeWork?.SWCosts, false, operationSign, scopeWork?.ChangeScopeWorkId);
+                            _scopeWork.UpdateParentCosts(parentContrId, scopeWork?.SWCosts, true, operationSign, scopeWork?.ChangeScopeWorkId);
+                        }
+                    }
+                    else
+                    {
+                        while (parentContrId != 0)
+                        {
+                            _scopeWork.UpdateParentCosts(parentContrId, scopeWork?.SWCosts, true, operationSign, scopeWork?.ChangeScopeWorkId);
+                            contract = parentContrId > 0 ? _contractService.GetById(parentContrId) : null;
+                            contractType = _contractService.GetContractType(contract, out parentContrId);
                         }
                     }
                 }
@@ -328,61 +320,31 @@ namespace MvcLayer.Controllers
         [Authorize(Policy = "DeletePolicy")]
         public async Task<IActionResult> ShowResultDelete(int? id)
         {
-            var scopeWork = _scopeWork.GetById((int)id);            
-            var contract =  _contractService.GetById((int)scopeWork.ContractId);
-
-            if (contract is not null)
+            var scopeWork = _scopeWork.GetById((int)id);
+            var swCosts = _swCostService.GetById((int)id);            
+            if (scopeWork is not null)
             {
-                int mainContractId = 0;
-
-                if (_contractService.IsNotGenContract(contract.Id, out mainContractId))
+                var contrId = scopeWork.ContractId;
+                int parentContrId = 0;
+                var contract = contrId.HasValue ? _contractService.GetById((int)contrId) : null;
+                var contractType = _contractService.GetContractType(contract, out parentContrId);
+                var parents = _contractService.GetParentsList(contract);
+                int oper = (contractType == ContractType.Agreement) || (contractType == ContractType.SubContract) ? 1 : -1;
+                foreach (var cost in scopeWork.SWCosts)
                 {
-                    foreach (var cost in scopeWork.SWCosts)
-                    {
-                        var costs = _mapper.Map<SWCostDTO>(cost);
-                        int ? mainContrScpId = 0;                        
-                        if (contract?.IsOneOfMultiple ?? false)
-                        {
-
-                            if (_contractService.IsThereScopeWorks(mainContractId, false, out mainContrScpId))
-                            {
-                                _scopeWork.RemoveOneCostOfMainContract(mainContrScpId, costs);
-                            }
-
-                            if (_contractService.IsThereScopeWorks(mainContractId, true, out mainContrScpId))
-                            {
-                                _scopeWork.RemoveOneCostOfMainContract(mainContrScpId, costs);
-                            }
-                        }
-                        else
-                        {
-                            _scopeWork.AddOrSubstractCostsOwnForceMnContract(mainContractId, new List<SWCostDTO> { costs }, 1);
-                        }
-                    }
+                    _scopeWork.RemoveSubContractCost((int)cost.Id, (int)contrId, parents, oper);
+                    _swCostService.Delete((int)cost.Id);
                 }
-                else
-                {
-                    foreach (var cost in scopeWork.SWCosts)
-                    {
-                        _scopeWork.RemoveExistOwnForce((int)scopeWork.Id, (int)cost.Id);
-                    }
-                }
+                //TODO: сделать проверку, есть ли дочерние договора, и есть ли стоимости у договора, для того чтобы удалить объем работ или не трогать
+                //var isLastSwCost = _swCostService.Find(x => x.ScopeWorkId == scpId).Count() > 0 ? false : true;
+                //if (isLastSwCost)
+                //{
+                //    _scopeWork.DeleteAllScopeWorkContract((int)scpId);
+
+                //}
             }
-
-
-            foreach (var cost in scopeWork.SWCosts)
-            {
-                _swCostService.Delete((int)cost.Id);
-            }            
-            var isLastSwCost = _swCostService.Find(x => x.ScopeWorkId == scopeWork.Id).Count() > 0 ? false : true;
-            if (isLastSwCost)
-            {
-                _scopeWork.DeleteAllScopeWorkContract((int)scopeWork.Id);
-
-            }
-
             ViewData["reload"] = "Yes";
-            return PartialView("_Message", new ModalViewModel("Запись успешно удалена.", "Результат удаления", "Хорошо"));
+            return PartialView("_Message", new ModalViewModel("Запись успешно удалена.", "Результат удаления", "Хорошо"));           
         }
 
         public IActionResult GetCostDeviation(string currentFilter, int? pageNum, string searchString)
@@ -790,28 +752,35 @@ namespace MvcLayer.Controllers
         {
             List<SWCostDTO> costs = new List<SWCostDTO>();
             costs.AddRange(_mapper.Map<List<SWCostDTO>>(model.SWCosts));
+            int parentContrId = 0;
             var contract = _contractService.GetById(contractId);
+            var contractType = _contractService.GetContractType(contract, out parentContrId);
 
-            if (contract is not null)
+            if (contractType == ContractType.GenСontract)
             {
-                int mainContractId = 0;
-                if (_contractService.IsNotGenContract(contract.Id, out mainContractId))
+                _scopeWork.UpdateParentCosts(contractId, costs, isOwnForces: true, operatorSign: 1, changeScopeId: model?.Id);
+            }
+            else if (contractType == ContractType.MultipleContract)
+            {
+                _scopeWork.UpdateParentCosts(parentContrId, costs, true, 1, model?.Id);
+                _scopeWork.UpdateParentCosts(parentContrId, costs, false, 1, model?.Id);
+
+            }
+            else
+            {
+                while (parentContrId != 0)
                 {
-                    if (!contract.IsOneOfMultiple)
-                    {
-                        _scopeWork.UpdateCostOwnForceMnContract(mainContractId, (int)model?.Id, costs);
-                    }
-                    else
-                    {
-                        _scopeWork.UpdateCostOwnForceMnContract(mainContractId, (int)model?.Id, costs, true);
-                    }
+                    _scopeWork.UpdateParentCosts(parentContrId, costs, true, -1, model?.Id);
+                    contract = parentContrId > 0 ? _contractService.GetById(parentContrId) : null;
+                    contractType = _contractService.GetContractType(contract, out parentContrId);
                 }
             }
+
             foreach (var swCost in model.SWCosts)
             {
                 _swCostService.Update(_mapper.Map<SWCostDTO>(swCost));
-            }
-            return RedirectToAction("GetByContractId", new { contractId = contractId, returnContractId = returnContractId });
+            }            
+            return RedirectToAction("GetByContractId", new { contractId = contractId, returnContractId = returnContractId });           
         }
 
         public IActionResult GetPeriodAmendment(int Id)
@@ -890,59 +859,59 @@ namespace MvcLayer.Controllers
         }  
 
         #region AdditionsMethods
-        private void CreateOrUpdateSWCostsOfMainContract(int multipleContractId, bool isOwnForses, List<SWCostDTO> costs)
-        {
-            int? scopeId = null;
-            var contractMultiple = _contractService.GetById(multipleContractId);
+        //private void CreateOrUpdateSWCostsOfMainContract(int multipleContractId, bool isOwnForses, List<SWCostDTO> costs)
+        //{
+        //    int? scopeId = null;
+        //    var contractMultiple = _contractService.GetById(multipleContractId);
 
-            //если есть объем работ у главного договора
-            if (_contractService.IsThereScopeWorks(contractMultiple.Id, isOwnForses, out scopeId))
-            {
-                //есть стоимости у гл.дог-ра (ищем по ID объема работ)                  
-                if (_contractService.IsThereSWCosts(scopeId))
-                {
-                    ///обновляем соответствующие периоды, добавляя из подобъекта, и если их нет создаем с этим же периодом
-                    _scopeWork.AddSWCostForMainContract(scopeId, costs);
-                }
-                else
-                {
-                    //создаем, копируя с подобъекта
-                    _scopeWork.CreateSWCostForMainContract(scopeId, costs, isOwnForses);
-                }
-            }
-            else
-            {
-                //создаем объем и копируем ему данные стоимости из подобъекта
-                var scopeWorkId = _scopeWork.Create(new ScopeWorkDTO
-                {
-                    ContractId = contractMultiple.Id,
-                    IsOwnForces = isOwnForses,
-                    IsChange = false
-                });
+        //    //если есть объем работ у главного договора
+        //    if (_contractService.IsThereScopeWorks(contractMultiple.Id, isOwnForses, out scopeId))
+        //    {
+        //        //есть стоимости у гл.дог-ра (ищем по ID объема работ)                  
+        //        if (_contractService.IsThereSWCosts(scopeId))
+        //        {
+        //            ///обновляем соответствующие периоды, добавляя из подобъекта, и если их нет создаем с этим же периодом
+        //            _scopeWork.AddSWCostForMainContract(scopeId, costs);
+        //        }
+        //        else
+        //        {
+        //            //создаем, копируя с подобъекта
+        //            _scopeWork.CreateSWCostForMainContract(scopeId, costs, isOwnForses);
+        //        }
+        //    }
+        //    else
+        //    {
+        //        //создаем объем и копируем ему данные стоимости из подобъекта
+        //        var scopeWorkId = _scopeWork.Create(new ScopeWorkDTO
+        //        {
+        //            ContractId = contractMultiple.Id,
+        //            IsOwnForces = isOwnForses,
+        //            IsChange = false
+        //        });
 
-                _scopeWork.CreateSWCostForMainContract(scopeWorkId, costs, isOwnForses);
-            }
-        }
+        //        _scopeWork.CreateSWCostForMainContract(scopeWorkId, costs, isOwnForses);
+        //    }
+        //}
 
-        /// <summary>
-        /// обновляем глав. договор, удалением старых и добовления новых стоимостей
-        /// </summary>
-        /// <param name="multipleContractId"></param>
-        /// <param name="changeScopeId"></param>
-        /// <param name="isOwnForses"></param>
-        /// <param name="costs"></param>
-        private void UpdateSWCostsOfMainContract(int multipleContractId, int changeScopeId, bool isOwnForses, List<SWCostDTO> costs)
-        {
-            int? scopeId = null;
-            var contractMultiple = _contractService.GetById(multipleContractId);
+        ///// <summary>
+        ///// обновляем глав. договор, удалением старых и добовления новых стоимостей
+        ///// </summary>
+        ///// <param name="multipleContractId"></param>
+        ///// <param name="changeScopeId"></param>
+        ///// <param name="isOwnForses"></param>
+        ///// <param name="costs"></param>
+        //private void UpdateSWCostsOfMainContract(int multipleContractId, int changeScopeId, bool isOwnForses, List<SWCostDTO> costs)
+        //{
+        //    int? scopeId = null;
+        //    var contractMultiple = _contractService.GetById(multipleContractId);
 
-            //если есть объем работ у главного договора, проверяем стоимости
-            if (_contractService.IsThereScopeWorks(contractMultiple.Id, isOwnForses, out scopeId))
-            {
-                ///обновляем соответствующие периоды, вычетаем из объема договора старый объем подобъекта и добавляем новый объем
-                _scopeWork.UpdateCostOwnForceMnContract(multipleContractId, changeScopeId, costs);
-            }
-        }
+        //    //если есть объем работ у главного договора, проверяем стоимости
+        //    if (_contractService.IsThereScopeWorks(contractMultiple.Id, isOwnForses, out scopeId))
+        //    {
+        //        ///обновляем соответствующие периоды, вычетаем из объема договора старый объем подобъекта и добавляем новый объем
+        //        _scopeWork.UpdateCostOwnForceMnContract(multipleContractId, changeScopeId, costs);
+        //    }
+        //}
         #endregion
     }
 }
